pyproject.toml
[project]
name = "pat-ai"
version = "0.1.0"
description = "Pat.AI/특허분석 AI - RAG 기반 특허 검색 시스템"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "python-multipart>=0.0.6",
    "jinja2>=3.1.2",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "sqlalchemy>=2.0.0",
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
    "pypdf>=3.17.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "aiofiles>=23.0.0",
    "pydantic-settings>=2.10.1",
]

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
